+++ models/predictor.py
@@ -19,7 +19,7 @@ def predict(matrix_U, matrix_V, topK, matrix_Train, bias=None, measure="Cosine",
         if len(vector_train.nonzero()[0]) > 0:
             vector_predict = sub_routine(vector_u, matrix_V, vector_train, bias, measure, topK=topK, gpu=gpu)
         else:
-            vector_predict = np.zeros(topK)
+            vector_predict = np.zeros(topK, dtype=np.float32)
 
         prediction.append(vector_predict)
 
@@ -48,7 +48,7 @@ def sub_routine(vector_u, matrix_V, vector_train, bias, measure, topK=500, gpu=F
         import cupy as cp
         candidate_index = cp.argpartition(-vector_predict, topK+len(train_index))[:topK+len(train_index)]
         vector_predict = candidate_index[vector_predict[candidate_index].argsort()[::-1]]
-        vector_predict = cp.asnumpy(vector_predict)
+        vector_predict = cp.asnumpy(vector_predict).astype(np.float32)
     else:
         candidate_index = np.argpartition(-vector_predict, topK+len(train_index))[:topK+len(train_index)]
         vector_predict = candidate_index[vector_predict[candidate_index].argsort()[::-1]]
